// next-ui/src/generated/openapi/types.gen.ts
export type FollowCreationDto = {
    seriesId?: string;
    title?: string;
    url: string;
};

export type FollowDto = {
    addedAt: Date;
    enabled: boolean;
    id: string;
    lastCheckedAt?: Date;
    libraryId: string;
    seriesId?: string;
    title?: string;
    url: string;
};

export type FollowScheduleDto = {
    checkTime?: string;
    enabled: boolean;
    intervalHours: number;
    lastCheckTime?: Date;
    libraryId: string;
    scheduleMode: string;
};

export type FollowScheduleUpdateDto = {
    checkTime?: string;
    enabled: boolean;
    intervalHours: number;
    scheduleMode: string;
};

export type FollowUpdateDto = {
    enabled?: boolean;
    title?: string;
};